refactor(common/Table): extract logic from DateFilter
