+++ src/module/car/service-log/ui/tables/service-logs/service-logs.tsx
@@ -6,9 +6,9 @@ import { useEffect, useMemo, useRef, useState } from 'react';
 
 import type { AuthIdentityPersistence } from '@/common/application/persistence-model/auth-identity';
 import { browserAuthClient } from '@/dependency/auth-client/browser';
+import { filterColumnByDate } from '@/lib/tanstack/table/filter';
 import type { ServiceLog } from '@/types';
 import { serviceCategoryMapping } from '@/types';
-import { filterColumnByDate } from '@/ui/table/compounds/date-filter/date-filter';
 import { Table } from '@/ui/table/table';
 import type { UserDto } from '@/user/application/dto/user';
 import { UserBadge } from '@/user/presentation/ui/badge/badge';
